move and rename sol into solana
